Rename components and css
